+++ internal/pkg/env/regex.go
@@ -18,8 +18,11 @@ package env
 import "regexp"
 
 var (
-	ReTTD    = regexp.MustCompile(`TTD=(.*)`)
-	ReCONFIG = regexp.MustCompile(`CONFIG_URL=(.*)`)
-	ReMEV    = regexp.MustCompile(`MEV=(.*)`)
-	ReXEEV   = regexp.MustCompile(`XEE_VERSION=(.*)`)
+	ReTTD       = regexp.MustCompile(`TTD=(.*)`)
+	ReCONFIG    = regexp.MustCompile(`CONFIG_URL=(.*)`)
+	ReGENESIS   = regexp.MustCompile(`GENESIS_URL=(.*)`)
+	ReDEPLOY    = regexp.MustCompile(`DEPLOY_BLOCK=(.*)`)
+	ReMEV       = regexp.MustCompile(`MEV=(.*)`)
+	ReXEEV      = regexp.MustCompile(`XEE_VERSION=(.*)`)
+	ReBOOTNODES = regexp.MustCompile(`CL_BOOTNODES=(.*)`)
 )