+++ internal/pkg/generate/types.go
@@ -60,8 +60,14 @@ type GenerationData struct {
 // DockerComposeData : Struct Data object to be applied to docker-compose script
 type DockerComposeData struct {
 	TTD                 bool
-	CcPrysmCfg          bool
-	VlPrysmCfg          bool
+	CcCustomCfg         bool
+	CcRemoteCfg         bool
+	CcRemoteGen         bool
+	CcRemoteDpl         bool
+	VlCustomCfg         bool
+	VlRemoteCfg         bool
+	VlRemoteGen         bool
+	VlRemoteDpl         bool
 	XeeVersion          bool
 	Mev                 bool
 	MevPort             string
@@ -81,6 +87,7 @@ type DockerComposeData struct {
 	ElExtraFlags        []string
 	ClExtraFlags        []string
 	VlExtraFlags        []string
+	Bootnodes           []string
 	MapAllPorts         bool
 	SplittedNetwork     bool
 }